35/Flask-SocketIO/requirements.txt
Flask==3.1.2
Flask-SocketIO==5.3.6
flask-security-too 
flask-sqlalchemy 
email-validator